implement `newpost` endpoint
it writes a new file, then redirects the user to edit it
