--- src/routes/newpost/+page.svelte
@@ -1,3 +1,0 @@
-<script lang="ts">
-  export let data;
-</script>
